requirements.txt
Flask==3.0.0
requests==2.31.0
python-whois==0.8.0
beautifulsoup4==4.12.2
reportlab==4.1.0
